requirements.txt
# requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
firebase-admin==6.1.0
pydantic==1.10.7
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6